Check for file existence before deletion (#22)
Check for file existence before deletion, to deal with the case that the
provisioner needs to be run multiple times in order to fully delete the
required users

```
TASK [users : Provision directory] *********************************************
ok: [tr-microk8s1]
fatal: [wato-bastion]: FAILED! => {"changed": true, "cmd": ["npx", "--yes", "@watonomous/linux-directory-provisioner@v0.0.5-alpha.3", "--no-confirm", "--config", "-"], "delta": "0:00:02.045768", "end": "2025-04-28 05:14:09.359677", "msg": "non-zero return code", "rc": 1, "start": "2025-04-28 05:14:07.313909", "stderr": "$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\n$ rm -f undefined\
```

---------

+++ src/index.mjs
@@ -16,6 +16,7 @@ import {
   objectMap,
   makeQuotaConfig,
   QUOTA_BLOCK_SIZE,
+  doesPathExist
 } from "./utils.mjs";
 import { validateConfig } from "./schema.mjs";
 
@@ -280,7 +281,11 @@ console.time("deleteSSHKeys");
 await Promise.all(
   usersToDelete.map(async (username) => {
     const sshAuthorizedKeysPath = configSSHAuthorizedKeysPathTemplate.replace(/%u/g, username).replace(/%U/g, users[username].uid);
-    await $`rm -f ${sshAuthorizedKeysPath}`;
+    if (await doesPathExist(sshAuthorizedKeysPath)) {
+      await $`rm -f ${sshAuthorizedKeysPath}`;
+    } else {
+      console.warn(`WARNING: No sshAuthorizedKeysPath for user ${username}`);
+    }
   })
 );
 console.timeLog("deleteSSHKeys");
@@ -289,12 +294,16 @@ console.timeLog("deleteSSHKeys");
 console.log(`Deleting managed user directories for ${usersToDelete.length} users...`);
 console.time("deleteManagedDirs");
 await Promise.all(
-  usersToDelete.map(async (u) => Promise.all(
-    config.managed_user_directories.map(async (d) => {
-      const formatdir = d.replace("%u", users[u].username).replace("%U", users[u].uid);
-      await $`rm -rf ${formatdir}`;
-    })
-  )
+    usersToDelete.map(async (u) => Promise.all(
+      config.managed_user_directories.map(async (d) => {
+        const formatdir = d.replace("%u", users[u].username).replace("%U", users[u].uid);
+        if (await doesPathExist(formatdir)) {
+          await $`rm -rf ${formatdir}`;
+        } else {
+          console.warn(`WARNING: Directory doesn't exist for user ${u}: ${formatdir}`);
+        }
+      })
+    )
   )
 )
 console.timeLog("deleteManagedDirs");
